fix for #9237
Appears to be an order problem with shutting down postgres when there
are active sessions/logins.

To reproduce (prior to fix).
  - start vm
  - login to web
  - immediately issue a shutdown after the page loads

This forces a dependency on postgres as well as omero for omero web to
ensure postgres is the last thing shutdown.  Debian's LSB initscripts
are run concurrently so a race condition can occur when they run at the
same time.